requirements.txt
celery==4.1.0
Django==1.11.5
django-celery-results==1.0.1
django-cleanup==1.0.1
django-extensions==1.9.1
django-materializecss-form==1.0.1
django-materialize-forms==0.2.0
gunicorn==19.7.1
junos-eznc==2.1.7
lxml==3.8.0
mod_wsgi==4.5.20
mod_wsgi-httpd
ncclient==0.5.3
paramiko==2.1.2
pylint==1.7.1
pyyaml==3.12
scp==0.10.2
Twisted==17.9.0
uwsgi==2.0.15
xmltodict==0.11.0



